chore: rename MCDALayerParameters component
