+++ src/features/mcda/components/MCDALayerEditor/MCDALayerParameters/constants.ts
